Make a stub 'fork' command that doesn't actually do anything yet.
Because it's a global command, not a guild command, it might take up to an hour for it to actually show up for use in guilds. I THINK it worked but can't tell yet because of that.

This will require re-authing the bot on your channel as it also requires application.command oauth scope.

Part of #34.