Changed how the bot decides if it's fetched back far enough in history.
Before, for when a bot was rebuilding IDF and there were a lot of messages, it was possible a new message could come in while the old history was being fetched. We'd see the new message, and think "OK we've got everything for this channel".

Now, the bot keeps track of messages it's seen that have come in live since the bot started, and doesn't stop fetchign history if it hits one of those.

Part of #35.