Radical overhaul to IDF/TFIDF system.
IDF is generated a cached when a guild is first seen. All that is kept is a by-stemmed-word index of which documents have at least one instance of each word.

When TFIDF is genrated, we fetch the messages for that channel on demand and then do the same calcualtion using the cached IDF.

IDF doesn't need to be regenerated all of that often; it will change relatively rarely for most guilds.

This change gets rid of most of the message caching logic which we'll want in some form again when we do #41, but don't need now (before this, it was possible to OOM if running for a long time).

This is a little slower when generating a suggested name, but not THAT slow (just the cost of fetching all messages in a channel sequentially), and it doesn't happen often.

Part of #35.