IDFs are now held around in memory, and when we note a forked message we don't persist the in-memory IDF.
In-memory IDFs might get kind of wonky (e.g. message edits, etc). The canonical way to generate them is to rebuild from scratch, which still happens on a timer.

We'll need the IDFs to be around and convienent to check on message reaction if there's a forked message to update.

Part of #34.