noteMessageIfFork now requests persistence.
Before, if a forked message came in after an IDF cache was created, and the bot was killed and restarted, then the bot wouldn't know about the forked message until the IDF cache was rebuilt.

Now it should be way less likely for the IDF Cache to not include MOST forked messages (although you could still have one come in in the moment the bot was down, which wouldn't be noticed until the next reindexing).

Part of #34.